Lab290 Left Shift Operator Program

+++ src/edu/javacore/Operateer_275/Lab290.java
@@ -2,7 +2,7 @@ package edu.javacore.Operateer_275;
 
 public class Lab290 {
 public static void main(String[] args) {
-	int a = 60;
+	long a = 60;
 	System.out.println(a << 32);
 	System.out.println(a << 33);
 	System.out.println(a << 64);
